+++ Packages/com.wrchat.woodon/Runtime/Core/CustomValue/Impl/MBool/Scripts/Core/MBool.cs
@@ -64,9 +64,9 @@ namespace WRC.Woodon
 			SendEvents();
 
 			if (Value == true)
-				SendEvents((int)MBoolEvent.OnTrue);
+				SendEvents(MBoolEvent.OnTrue);
 			else
-				SendEvents((int)MBoolEvent.OnFalse);
+				SendEvents(MBoolEvent.OnFalse);
 		}
 
 		public virtual void SetValue(bool newValue, bool isReciever = false)
